backend/generate_mcp_endpoints_tools/requirements.txt
PyYAML>=6.0